// src/core/types/entities/iron-lake/parameter/parameter-ehms/UpsertItem.ts
export type UpsertItem = {
    ParameterEhmsId: number
    CbmParameterId: string
    ComponentId: string
    CbmGroup: string
    CbmArea: string
    Model: string
    PsType: string
    TaskNumber: string
    DetailNumber: string
    CbmParameter: string
    Parameter: string
    TypeParameterId: string
    ValueMin: string
    ValueMax: string
    Uom: string
    StatusConverter: string
    StatusConverterDescription: string
    Status: string
    StatusDescription: string
    StartDate: string
    EndDate: string
    IsActive: boolean
}